+++ client/game/Empire/Planets/Resources/Resources.vue
@@ -3,7 +3,7 @@
      * Planetary Resources
      * this component lists all planetary resource types - min 0, max 4 (lucky tombworld!)
      ******************************************************************************************************************/
-    import Resource from "./Resource.vue";
+    import ResourceType from "./ResourceType.vue";
     export default {
         props: {
             resources: {
@@ -13,10 +13,14 @@
             planetid: {
                 type: String,
                 required: true
+            },
+            planetName: {
+                type: String,
+                required: true
             }
         },
         components: {
-            "resource": Resource
+            "resource-type": ResourceType
         }
     };
 </script>
@@ -24,11 +28,12 @@
 <template>
     <ul class="resources">
         <li v-for="resource in resources">
-            <resource :id="resource.id"
-                      :resourceType="resource.resourceType"
-                      :slots="resource.slots"
-                      :harvesters="resource.harvesters"
-                      :planetid="planetid"/>
+            <resource-type :id="resource.id"
+                           :resourceType="resource.resourceType"
+                           :slots="resource.slots"
+                           :harvesters="resource.harvesters"
+                           :planetid="planetid"
+                           :planetName="planetName" />
         </li>
     </ul>
 </template>