+++ client/game/store/actions.js
@@ -9,7 +9,6 @@ import axios from "axios";
 import cfg from "../../config";
 
 const ACTIONS = {
-
     /* =================================================================================================================
      * COMMON ACTIONS===================================================================================================
      * ============================================================================================================== */
@@ -68,6 +67,41 @@ const ACTIONS = {
                 console.error(error);
                 ctx.commit("SAVING_STAR_NAME", {id: payload.id, saving: false});
                 ctx.commit("EDITING_STAR_NAME", {id: payload.id, editing: false});
+            });
+    },
+
+    /*
+     * post "install harvester on planet" request
+     */
+    INSTALL_HARVESTER: function(ctx, payload) {
+        cfg.DEBUG && console.log("requesting install harvestester", payload);
+        ctx.commit("SAVING_INSTALL_HARVESTER", {resourceId: payload.resourceId, saving: true});
+        axios
+            .post("/api/game/empire/harvester/install", payload)
+            .then(response => {
+                if (response.status === 200 && response.data) {
+                    ctx.commit("ADD_HARVESTER", {
+                        harvesterType: payload.harvesterType,
+                        planet: payload.planet,
+                        id: response.data.harvester.id,
+                        turnsUntilComplete: response.data.harvester.turnsUntilComplete,
+                        isHarvesting: false
+                    });
+
+                } else if (response.data.error) {
+                    console.error(response.data.error);
+                }
+                ctx.commit("SAVING_INSTALL_HARVESTER", {
+                    resourceId: payload.resourceId,
+                    saving: false
+                });
+            })
+            .catch(error => {
+                console.error(error);
+                ctx.commit("SAVING_INSTALL_HARVESTER", {
+                    resourceId: payload.resourceId,
+                    saving: false
+                });
             });
     }
 };